+++ lib/index.js
@@ -4,20 +4,20 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 
-var _rxToMobservable = require('./rxToMobservable');
+var _rxToMobx = require('./rxToMobx');
 
-Object.defineProperty(exports, 'rxToMobservable', {
+Object.defineProperty(exports, 'rxToMobx', {
   enumerable: true,
   get: function get() {
-    return _rxToMobservable.rxToMobservable;
+    return _rxToMobx.rxToMobx;
   }
 });
 
-var _mobservableToRx = require('./mobservableToRx');
+var _mobxToRx = require('./mobxToRx');
 
-Object.defineProperty(exports, 'mobservableToRx', {
+Object.defineProperty(exports, 'mobxToRx', {
   enumerable: true,
   get: function get() {
-    return _mobservableToRx.mobservableToRx;
+    return _mobxToRx.mobxToRx;
   }
 });
